+++ apps/wallet-mobile/src/features/Notifications/useCases/common/hooks.ts
@@ -5,7 +5,7 @@ import {PermissionsAndroid} from 'react-native'
 
 import {notificationManager} from './notification-manager'
 import {parseNotificationId} from './notifications'
-import {usePrimaryTokenPriceChangedNotification} from './primary-token-price-updated-notification'
+import {usePrimaryTokenPriceChangedNotification} from './primary-token-price-changed-notification'
 import {useTransactionReceivedNotifications} from './transaction-received-notification'
 
 let initialized = false
