+++ apps/wallet-mobile/src/features/Notifications/useCases/common/notifications.ts
@@ -1,6 +1,7 @@
 import {Notification, Notifications} from '@jamsinclair/react-native-notifications'
 import {mountAsyncStorage} from '@yoroi/common'
 import {Notifications as NotificationTypes} from '@yoroi/types'
+
 import {formatCurrency, getCurrencySymbol} from '../../../Settings/useCases/changeAppSettings/Currency/CurrencyContext'
 
 export const generateNotificationId = (): number => {
